Initial checkin of App Code
This commit begins the structure of the sinatra app.
Tests are placed in the /spec folder, which test the
app contained in the /app folder.

The initial REST API is written for the book, creator,
and publisher roles, as well as the Sinatra web app to
present the JSON endpoints.